Extract DeviceType to its own class
